Fix: mark from DATA and BSS sections
Constants are initialized into the DATA and BSS sections, so the
collector must mark from these roots, in addition to fiber stacks.

Since we can't access the address of an extern, we rely on a small
C helper script to access some linker symbols (`__data_start`,
`__bss_start` and `_end`).