Fix: replace recursive marking with a LIFO stack
Programs that keep deeply nested tree of objects used to cause stack
overflows, because the collector was using recursive calls to the
Collector_markRegion function. The Crystal compiler is subject to
this issue, when compiling large programs, such as itself for
example.

This patch fixes the issue by replacing the recursive markRegion
function with a loop and a manual stack of roots to trace.

The `GC_mark_region` function has been replaced with the
`GC_add_roots` function with the same signature. The function now
pushes a stack root trace, which will be marked later, instead of
being marked immediately.