Fix: move finalizer when reallocating object
Reallocating an object should move any finalizer set on the object.
Keeping the finalizer on the old object could lead to finalizing an
old reference despite a newer reference being still alive.

This shouldn't happen in Crystal, since references are unexpected to
be reallocated, but still implement to avoid future issues.