Register finalizers in hash table
Finalizers aren't common enough that they should require a void*
overhead on each and every object (4 to 8 bytes per individual
allocation) and to iterate the whole small and large HEAP for
unmarked objects with finalizers.

This patch moves finalizers to a hash table (key=object,
value=callback) attached to the global allocator, and registers
finalizers for both small and large objects. The collector
merely iterates the hash for unmarked objects to finalize.