Fix: delay Fiber spawn to Crystal runtime startup
We can't spawn a fiber right into `GC.init` because this method is
called before anything else from the crystal runtime, and spawning a
fiber needs both the GC and the runtime to be available.

The solution is simple: just use a global class variable to hold the
fiber!

Also fixes an issue where `GC.collect` may be called from the
current Fiber as well as the Signal loop fiber (maybe other fibers),
which would end up having an overwritten `@@pending` fiber, and thus
no fiber to resume to. This is fixed by using, you guess it: a Mutex.