Implement missing GC methods to work properly with Crystal 1.0+
At some point during the development of Crystal, the responsibility of pushing
the GC roots was removed from `fiber.cr` into the default GC implementation
`bohem.cr`. Thus this code was missing when compiling with `-Dgc_none`. Add a
`before_collect` callback to push the fiber stacks as GC roots and adapt to
changes to the methods to return the stack bottom of the current thread.