Fix: protect GC.collect using an Atomic::Flag
A situation in Crystal causes the Signal Loop Fiber to eventually
call `GC.collect` immediately after the main Fiber calling
`GC.collect` which results in a deadlock (`@@pending` is overwritten
and the main fiber is never resumed).

We can't use a Mutex to fix the issue (it allocates memory), and
neither can we use a Thread::Mutext because it may allocate memory
(it may raise exceptions) and using pthreads directly would either
check for reentrancy and either crash (ERRORCHECK) or fail
(RECURSIVE) because the same thread may call the `GC.collect`
function.